Refactoring: Class -> Functional
